WIP: Create temporary files in a safer way
Previously either a temporary file was created and then removed
or just a temporary file name was created.  Both of these are
unsafe in two ways.

Another program could create a file of the same name in the
meantime either by accident or on purpose and trick ezQuake
into operating on that.  Or ezQuake would overwrite another
application's temporary data, which happened to pick the same
temporary file name.

Now both temporary files and directories are created and ensured
to exist.  No other well behaved program should interfere with
ezQuake's temporary data and the other way around.

It's still possible for a malicious program to replace the
temporary file and have ezQuake operate on it.  To prevent that
all temporary file and file system functions would have to be
rewritten to deal with `FILE *` instead of file paths.

To improve clarity the type of several function arguments has
been changed from `char *` to `const char *`.

Fixes #223.